requirements.txt
Django==3.0.5
django-model-utils==4.0.0
django-mptt==0.11.0
django-allauth==0.41.0
PyJWT==1.7.1
celery==4.4.2
easy-thumbnails==2.7
